remove traits
